backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
cryptography
requests
